Improve the efficiency of the get_shipments call
After consulting with the ShipStation developers they informed us the CreateDate for Shipment could be used as a proxy for ShipDate which does not contain time information.  Now we can return just the shipments that have shipped since we last checked which also allows us to comply with the new API limits set by ShipStation.

[Fixes IN-306]